requirements.txt
# Core dependencies for Consciousness Framework
openai>=1.0.0
anthropic>=0.7.0
aiohttp>=3.8.0
requests>=2.28.0
PyYAML>=6.0
python-dotenv>=1.0.0
